+++ samples/backtest_bbands.py
@@ -28,6 +28,7 @@ from basana.core.logs import StructuredMessage
 from basana.external.binance import csv
 import basana as bs
 import basana.backtesting.exchange as backtesting_exchange
+import basana.backtesting.fees as backtesting_fees
 
 from samples.backtesting import position_manager
 from samples.strategies import bbands
@@ -42,15 +43,19 @@ async def main():
     stop_loss_pct = Decimal(5)
 
     # We'll be opening short positions so we need to set a lending strategy when initializing the exchange.
-    lending_strategy = lending.MarginLoans(pair.quote_symbol, default_conditions=lending.MarginLoanConditions(
-        interest_symbol=pair.quote_symbol, interest_percentage=Decimal("7"),
-        interest_period=datetime.timedelta(days=365), min_interest=Decimal("0.01"),
-        margin_requirement=Decimal("0.5")
-    ))
+    lending_strategy = lending.MarginLoans(
+        pair.quote_symbol, Decimal("0.5"),
+        default_conditions=lending.MarginLoanConditions(
+            interest_symbol=pair.quote_symbol, interest_percentage=Decimal("7"),
+            interest_period=datetime.timedelta(days=365), min_interest=Decimal("0.01"),
+        )
+    )
     exchange = backtesting_exchange.Exchange(
         event_dispatcher,
         initial_balances={pair.quote_symbol: Decimal(1200)},
+        fee_strategy=backtesting_fees.Percentage(percentage=Decimal("0.1"), min_fee=Decimal("0.01")),
         lending_strategy=lending_strategy,
+        immediate_order_processing=True
     )
     exchange.set_symbol_precision(pair.base_symbol, 8)
     exchange.set_symbol_precision(pair.quote_symbol, 2)