+++ samples/backtest_sma.py
@@ -40,6 +40,7 @@ async def main():
     exchange = backtesting_exchange.Exchange(
         event_dispatcher,
         initial_balances={pair.quote_symbol: Decimal(1200)},
+        immediate_order_processing=True
     )
     exchange.set_symbol_precision(pair.base_symbol, 8)
     exchange.set_symbol_precision(pair.quote_symbol, 2)