+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.8
+
+### Features
+
+* Backtesting `OrderInfo` now includes pair and order fills.
+* Backtesting exchange returns a full `OrderInfo` when creating an order.
+* Backtesting exchange support for immediate order processing.
+
+### Bug fixes
+
+* Backtesting margin level calculations were wrong. `margin_requirement` was moved from `MarginLoanConditions` into `MarginLoans`
+
 ## 1.7.1
 
 ### Bug fixes
@@ -47,7 +59,7 @@
 
 ## 1.5.0
 
-* `basana.backtesting.fees.Percentage` now supports a minimum fee.
+* `backtesting.fees.Percentage` now supports a minimum fee.
 
 ## 1.4.1